+++ src/gamspy/_algebra/operation.py
@@ -3,9 +3,11 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import gamspy._algebra.condition as condition
+import gamspy._algebra.domain as domain
 import gamspy._algebra.expression as expression
 import gamspy._algebra.operable as operable
 import gamspy._symbols as syms
+import gamspy._symbols.implicits as implicits
 import gamspy.utils as utils
 from gamspy.exceptions import ValidationError
 
@@ -13,26 +15,79 @@ if TYPE_CHECKING:
     from gams.transfer import Alias, Parameter, Set
 
     from gamspy._algebra import Domain
+    from gamspy._algebra.condition import Condition
     from gamspy._algebra.expression import Expression
-    from gamspy._symbols.implicits import ImplicitParameter, ImplicitVariable
+    from gamspy._symbols.implicits import (
+        ImplicitParameter,
+        ImplicitSet,
+        ImplicitVariable,
+    )
 
 
 class Operation(operable.Operable):
     def __init__(
         self,
-        domain: Set | Alias | tuple[Set | Alias] | Domain | Expression,
-        expression: (
-            Expression | ImplicitVariable | ImplicitParameter | int | bool
+        domain: Set
+        | Alias
+        | ImplicitSet
+        | tuple[Set | Alias]
+        | Domain
+        | Condition,
+        rhs: (
+            Expression
+            | Operation
+            | ImplicitVariable
+            | ImplicitParameter
+            | int
+            | bool
         ),
         op_name: str,
     ):
         self.domain = utils._to_list(domain)
         assert len(self.domain) > 0, "Operation requires at least one index"
-        self.expression = expression
+        self.rhs = rhs
         self._op_name = op_name
+        self.raw_domain = self.get_raw_domain()
 
         # allow conditions
         self.where = condition.Condition(self)
+
+    def get_raw_domain(self) -> list[Set | Alias | ImplicitSet]:
+        raw_domain = []
+        for elem in self.domain:
+            if isinstance(elem, condition.Condition):
+                if isinstance(elem.conditioning_on, implicits.ImplicitSet):
+                    raw_domain.append(elem.conditioning_on.parent)
+                elif isinstance(elem.conditioning_on, (syms.Set, syms.Alias)):
+                    raw_domain.append(elem.conditioning_on)
+                elif isinstance(elem.conditioning_on, domain.Domain):
+                    raw_domain += elem.conditioning_on.sets
+            elif isinstance(elem, domain.Domain):
+                raw_domain += elem.sets
+            elif isinstance(elem, implicits.ImplicitSet):
+                raw_domain.append(elem)
+            else:
+                raw_domain.append(elem)
+
+        return raw_domain
+
+    def validate_operation(self, control_stack):
+        for elem in self.raw_domain:
+            if isinstance(elem, implicits.ImplicitSet):
+                control_stack += [
+                    member
+                    for member in elem.domain
+                    if member not in control_stack
+                ]
+
+            if elem in control_stack:
+                raise ValidationError(f"Set {elem} is already in control")
+
+        stack = control_stack + self.raw_domain
+        if isinstance(self.rhs, expression.Expression):
+            self.rhs._validate_definition(stack)
+        elif isinstance(self.rhs, Operation):
+            self.rhs.validate_operation(stack)
 
     def _get_index_str(self) -> str:
         if len(self.domain) == 1:
@@ -74,18 +129,18 @@ class Operation(operable.Operable):
         output += index_str
         output += ","
 
-        if isinstance(self.expression, float):
-            self.expression = utils._map_special_values(self.expression)
+        if isinstance(self.rhs, float):
+            self.rhs = utils._map_special_values(self.rhs)
 
-        if isinstance(self.expression, bool):
-            self.expression = (
-                "yes" if self.expression is True else "no"  # type: ignore
+        if isinstance(self.rhs, bool):
+            self.rhs = (
+                "yes" if self.rhs is True else "no"  # type: ignore
             )
 
         expression_str = (
-            str(self.expression)
-            if isinstance(self.expression, (int, str))
-            else self.expression.gamsRepr()
+            str(self.rhs)
+            if isinstance(self.rhs, (int, str))
+            else self.rhs.gamsRepr()
         )
 
         output += expression_str
@@ -120,9 +175,9 @@ class Operation(operable.Operable):
             index_str += " ~ | ~ " + given_condition.latexRepr()
 
         expression_str = (
-            str(self.expression)
-            if isinstance(self.expression, (int, str))
-            else self.expression.latexRepr()
+            str(self.rhs)
+            if isinstance(self.rhs, (int, str))
+            else self.rhs.latexRepr()
         )
         representation = f"\\displaystyle \\{op_map[self._op_name]}_{{{index_str}}} {expression_str}"
         return representation