+++ src/gamspy/_symbols/implicits/implicit_parameter.py
@@ -64,6 +64,8 @@ class ImplicitParameter(ImplicitSymbol, operable.Operable):
             rhs,
         )
 
+        statement._validate_definition(domain)
+
         self.container._add_statement(statement)
         self.parent._assignment = statement