+++ CHANGELOG.md
@@ -6,11 +6,13 @@ GAMSPy 0.13.7
 - General
   - Support .where syntax for Card and Ord.
   - Return condition on where operations on the right instead of expression.
+  - Catch set is already under control errors early.
 - Documentation
   - Fix docstirng of the Card operation.
   - Add warning about non-professional licenses in addGamsCode docstring.
 - Testing
   - Add tests for .where syntax for Card and Ord.
+  - Add tests to catch set is already under control errors.
 
 GAMSPy 0.13.6
 -------------